@N5N3 2024-10-07
@nanosoldier `runtests(["LinearRationalExpectations", "EarlyStopping", "QuantumAlgebra", "QuantumPropagators", "Permanents", "FindMinimaxPolynomial", "CTFlows", "OptimalBids", "SliceSampling", "FSimBase", "DiffEqFinancial", "PlantBiophysics", "MLJTestIntegration", "CellularPotts", "CirculatorySystemModels", "IterativeLearningControl"])`