@KristofferC 2024-10-08
@charleskawczynski, you can get the built Julia from this PR from buildkite to easily test it yourself, for example:

![image](https://github.com/user-attachments/assets/c33049b7-99f9-4363-b9c4-f594ce9696db)