@N5N3 2024-10-08
@charleskawczynski I tried it, but the error print is still quite slow.