@charleskawczynski 2024-10-08
> @charleskawczynski, you can get the built Julia from this PR from buildkite to easily test it yourself, for example: ...

Ah, cool! I just tried downloading it but I'm getting

<img width="252" alt="Screenshot 2024-10-08 at 11 03 55 AM" src="https://github.com/user-attachments/assets/53cb13d6-b2b8-460b-804e-0f7aa44b9efe">

when I try to run `julia-dc689fe870/bin/julia`. Am I supposed to run it some other way?