@charleskawczynski 2024-10-11
If I try to run it from terminal, I get
```
/Users/charliek/Downloads/julia-dc689fe870Build/bin/julia 
dyld[65179]: Library not loaded: @rpath/libjulia.1.12.dylib
  Referenced from: <C0E46333-5FAF-3210-A327-5E41CB0C8AFE> /Users/charliek/Downloads/julia-dc689fe870Build/bin/julia
  Reason: tried: '/Users/charliek/Downloads/julia-dc689fe870Build/lib/libjulia.1.12.dylib' (code signature in <845EDEB7-45DE-332E-B6A8-5C21830FD847> '/Users/charliek/Downloads/julia-dc689fe870Build/lib/libjulia.1.12.0.dylib' not valid for use in process: library load disallowed by system policy), '/Users/charliek/Downloads/julia-dc689fe870Build/lib/libjulia.1.12.dylib' (code signature in <845EDEB7-45DE-332E-B6A8-5C21830FD847> '/Users/charliek/Downloads/julia-dc689fe870Build/lib/libjulia.1.12.0.dylib' not valid for use in process: library load disallowed by system policy), '/usr/local/lib/libjulia.1.12.dylib' (no such file), '/usr/lib/libjulia.1.12.dylib' (no such file, not in dyld cache)
Abort trap: 6
```